final website: Added twine div (tw-story) inside the right side div, and some data handling on the tabs (as a test, for all except the phone tab) by using onclick() functions and calls to the jQueryAjax.js script.
Situation: Needed to figure out a way to append the twine story on the website and start data handling.
Changes: Added twine div (tw-story) inside the right side div, and some data handling on the tabs (as a test, for all except the phone tab) by using onclick() functions and calls to the jQueryAjax.js script.